+++ src/extension.ts
@@ -11,6 +11,7 @@ import {
   setAppRunningStateChangeCallback,
   stopApp,
 } from "./run";
+import { setRunFromOverride } from "./set-run-from-override-command";
 import {
   shinyliveCreateFromActiveEditor,
   shinyliveCreateFromExplorer,
@@ -24,6 +25,10 @@ export async function activate(context: vscode.ExtensionContext) {
     vscode.commands.registerCommand("shiny.python.debugApp", pyDebugApp),
     vscode.commands.registerCommand("shiny.r.runApp", rRunApp),
     vscode.commands.registerCommand("shiny.stopApp", stopApp),
+    vscode.commands.registerCommand(
+      "shiny.setRunFromOverride",
+      setRunFromOverride
+    ),
     vscode.commands.registerCommand(
       "shiny.shinylive.createFromActiveEditor",
       shinyliveCreateFromActiveEditor