+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-New `shiny.timeoutOpenBrowser` option to increase default timeout.
+## [UNRELEASED]
+
+New `shiny.timeoutOpenBrowser` option sets the maximum time to wait for a Shiny app to be ready before attempting to open it in the browser. (@olivroy #82)
 
 ## 1.1.0